Rename app and bundle
