+++ android/app/BUCK
@@ -35,12 +35,12 @@ android_library(
 
 android_build_config(
     name = "build_config",
-    package = "com.darrektdev.sage",
+    package = "com.mindmymind.within",
 )
 
 android_resource(
     name = "res",
-    package = "com.darrektdev.sage",
+    package = "com.mindmymind.within",
     res = "src/main/res",
 )
