+++ android/app/src/main/java/com/mindmymind/within/DnDMode.java
@@ -1,4 +1,4 @@
-package com.darrektdev.sage;
+package com.mindmymind.within;
 import com.facebook.react.bridge.NativeModule;
 import com.facebook.react.bridge.ReactApplicationContext;
 import com.facebook.react.bridge.ReactContext;
