+++ android/app/src/main/java/com/mindmymind/within/MainActivity.java
@@ -1,4 +1,4 @@
-package com.darrektdev.sage;
+package com.mindmymind.within;
 
 import android.os.Bundle;
