+++ android/app/src/main/java/com/mindmymind/within/MainApplication.java
@@ -1,4 +1,4 @@
-package com.darrektdev.sage;
+package com.mindmymind.within;
 
 import android.app.Application;
 import android.content.Context;
@@ -12,7 +12,7 @@ import com.facebook.react.ReactNativeHost;
 import com.facebook.react.ReactPackage;
 import com.facebook.react.shell.MainReactPackage;
 import com.facebook.soloader.SoLoader;
-import com.darrektdev.sage.generated.BasePackageList;
+import com.mindmymind.within.generated.BasePackageList;
 
 import org.unimodules.adapters.react.ReactAdapterPackage;
 import org.unimodules.adapters.react.ModuleRegistryAdapter;
@@ -104,7 +104,7 @@ public class MainApplication extends Application implements ReactApplication {
          We use reflection here to pick up the class that initializes Flipper,
         since Flipper library is not available in release mode
         */
-        Class<?> aClass = Class.forName("com.darrektdev.sage.ReactNativeFlipper");
+        Class<?> aClass = Class.forName("com.mindmymind.within.ReactNativeFlipper");
         aClass
             .getMethod("initializeFlipper", Context.class, ReactInstanceManager.class)
             .invoke(null, context, reactInstanceManager);
