+++ android/app/src/main/java/com/mindmymind/within/MyAppPackage.java
@@ -1,4 +1,4 @@
-package com.darrektdev.sage;
+package com.mindmymind.within;
 import com.facebook.react.ReactPackage;
 import com.facebook.react.bridge.NativeModule;
 import com.facebook.react.bridge.ReactApplicationContext;
