+++ android/settings.gradle
@@ -1,4 +1,4 @@
-rootProject.name = 'Sage'
+rootProject.name = 'Within'
 include ':react-native-linear-gradient'
 project(':react-native-linear-gradient').projectDir = new File(rootProject.projectDir, '../node_modules/react-native-linear-gradient/android')
