+++ ios/Podfile
@@ -4,7 +4,7 @@ require_relative '../node_modules/@react-native-community/cli-platform-ios/nativ
 
 platform :ios, '11.0'
 
-target 'Sage' do
+target 'Within' do
   use_unimodules!
   config = use_native_modules!
