+++ ios/Within.xcodeproj/xcshareddata/xcschemes/Within.xcscheme
@@ -15,9 +15,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-               BuildableName = "Sage.app"
-               BlueprintName = "Sage"
-               ReferencedContainer = "container:Sage.xcodeproj">
+               BuildableName = "Within.app"
+               BlueprintName = "Within"
+               ReferencedContainer = "container:Within.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
       </BuildActionEntries>
@@ -33,9 +33,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "00E356ED1AD99517003FC87E"
-               BuildableName = "SageTests.xctest"
-               BlueprintName = "SageTests"
-               ReferencedContainer = "container:Sage.xcodeproj">
+               BuildableName = "WithinTests.xctest"
+               BlueprintName = "WithinTests"
+               ReferencedContainer = "container:Within.xcodeproj">
             </BuildableReference>
          </TestableReference>
       </Testables>
@@ -55,9 +55,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-            BuildableName = "Sage.app"
-            BlueprintName = "Sage"
-            ReferencedContainer = "container:Sage.xcodeproj">
+            BuildableName = "Within.app"
+            BlueprintName = "Within"
+            ReferencedContainer = "container:Within.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
    </LaunchAction>
@@ -72,9 +72,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-            BuildableName = "Sage.app"
-            BlueprintName = "Sage"
-            ReferencedContainer = "container:Sage.xcodeproj">
+            BuildableName = "Within.app"
+            BlueprintName = "Within"
+            ReferencedContainer = "container:Within.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
    </ProfileAction>
