+++ ios/Within/Images.xcassets/AppIcon.appiconset/Contents.json
