+++ ios/Within/Info.plist
@@ -25,7 +25,7 @@
 		<dict>
 			<key>CFBundleURLSchemes</key>
 			<array>
-				<string>com.darrekt.Sage</string>
+				<string>com.mindmymind.within</string>
 			</array>
 		</dict>
 	</array>
